Fix SymPolyMat.__call__ and Diabatizer.optimize
The evaluation of a SymPolyMat at a point x is now fixed (previously the
off-diagonal terms were half the expected value).
Several issues in Diabatizer are corrected so that optimizations get to
completion (at least in 1D for the moment...).
First test case using LiF 1D data is included in the TestDiabatizer
class (and has helped in finding spotting the bugs).
The fit result is correct and identical to preliminary tests.