Remove default constant term addition in NdPoly
Now the constant term is not added by default in a polynomial
instanciated by calling the constructor, i.e. whose terms are
given like P = NdPoly({(1,2,3): 4, ...)} without specifying the
(0,0,0) key. Fixes are provided to overcome this, in particular when
adding a constant to such a polynomial, i.e. P + 3.14 now works by
adding NdPoly({(0,0,0): 3.14}).

Meanwhile, the 2d2s Diabatizer test is now in a preliminary form and
has to be completed with assert calls. An extra pytest option should be
added if one want the plots to be shown (instead of popping up by
default).