Define SymPolyMat.__iter__, SymPolyMat.zero_like
The polynomials are now stored in a single list (and not a list of
lists) so that SymPolyMat.__iter__() is simply defined as this list,
i.e. it returns self._polys.__iter__().

Because of this change in self._polys, the (i,j) indexing had to be
corrected (a double index is flattened in row-major format, adaptated
to symmetric matrices).

Now SymPolyMat can be iterated through by elements, e.g.:
for w in SymPolyMat(3,3): ...

Addition of a convenience static method SymPolyMat.zero_like().