feat: Multi-domain diabatization
The Diabatizer class now can store several diabatic potential matrices,
each assigned to a specific domain and selected states in each domain.
This selection is made more rigorous by collecting domains as dicts
mapping a unique id to an array of points or energies, as well as
defining a domainMap which assigns a matrix to a domain id and its
selected states.
The mapping is modified via Diabatizer.setFitDomain().
The code and tests are adapted as little as possible so as to take the
new changes into account and work as expected. The SigleDiabatizer
class, which inherits from the new version of Diabatizer, is created
in order to work like the former before the changes.