fix: evaluate empty polynomial to zero
Previously, when calling a polynomial on a given point x, the returned
value P was initialized to 0 (int) and monomial to 1 (int) which were
implicitly casted to float and operations with them were broadcasted to
fit the size of x. This was giving problems when evaluating an empty
polynomial because P.size is eventually invoked, which is not allowed
here as P remains the initialized integer.
The issue is resoved by initializing P and monomial with the proper
size with zeros() and ones().
TestPoly.test_Empty() is updated to check this fix.