feat: SymPolyMat elements have distinct x0
The origin/shift in each of the matrix elements in a SymPolyMat can now
be distinct (setting is possible by accessing directly the elements,
with W[i,j].x0 = point_ij).
Setting a common origin can be done with the method set_common_x0(point).
Getting all of the origins it done via get_all_x0() which returns a dict
of the form {(i,j): point, ...}.

Tests of the SymPolyMat class are added to check basic functionning.
They pass.
Diabatizer is updated in order to accommodate those changes, but tests
must be updated too.