feat+fix: NdPoly up to maxdegree new and edit
The previous version of NdPoly.zero_maxdegree was only including terms
of the given degree and not those of lower degree. Fixed.

It contained a nested function get_tuples which is now made static and
private in the NdPoly class.

This can thus be used in the grow_degree method, which does the same as
the static method above, but adds terms to the calling instance and
leaves untouched previously set terms.

Corresponding tests are added and pass.