improve: degree of polynomial
Previously, the degree of a polynomial was stored as an attribute which
was updated at each coefficient setting.
This is now changed, as the true degree is evaluated and returned when
calling the degree property. By true we mean that zero terms are not
taken into account, e.g. 0*x^3 + x^2 is of degree 2 and not 3.

The behaviour is tested throughout the TestPoly class.