feat: Diabatizer residual weighting functions
Diabatizer class.
Although not used yet in the cost function calculation, the attribute
self._weight is defined to store a callable.
This is expected to return a value between 0 and 1, taking an energy
value as the argument.
Some predefined weighting functions are added, based on Gaussian and
expenential decaying functions.