feat: new DampedSymPolyMat class
Definition of a new class DampedSymPolyMat, inheriting from SymPolyMat.
It contains an additional triangular "matrix" of callables.
These can be anything, preferably damping functions, which can be used
to weight off-diagonal terms in the SymPolyMat.
Convenience classes Gaussian and Lorenzian (corrected versions of the
old GaussianDamping and LorenzianDamping) are provided for this
purpose.