refactor(Diabatizer): minimize instead of lst_sq
Define a cost function as the weighted RMSE and minimize it with a
scipy.optimize.minimize routine. This replaces the former least squares
minimization which was not flexible enough for weighting the points.

Store weights explicitely as arrays of the same size of the energies in
each domain.
Reset (resp. delete) them when domain is turned to automatic weighting
(resp. removed).

Rename, refactor and redocument methods handling diabatic
matrices deconstructtion and reconstruction.