refactor(Diabatizer): list of states by domain
Make self._domain_map a list of dicts instead of dict of dicts.
Rename it self._states_by_domain.

The first level dict was merely mapping a matrix id (i.e. an integer)
to a dict of domain_id -> (s1, s2, ...).
Since throughout the class we are iterating through martrices ids, there
is no need for a dict.