fix: decode NdPoly nested in SymPolymat
The JSON parser's object_hook transforms the most deeply nested JSON
objects into Python dicts. Thus When parsing at the level of SymPolyMat,
the contained NdPolys are already constructed.
Thus, no need of reparsing those, otherwise we gat an error.

This is potentially dangerous, as this correction relies on a subtle
undocumented behaviour of the standard JSON decoder.
We keep it like this for the moment...