refactor: move JSON de/encoding to methods
In order to have major control on how JSON encoding and decoding of
objects is carried out, JSON related functionalities are moved into the
concerned classes.

First, we now make the disticntion between:
(1) JSON strings (e.g. as written in files)
(2) Python representation of a JSON object as a dict, aka JSON_dict
Writing and reading to/from JSON files/strings (1) is an operation handled
by the standard json module so as to dump/retreive the content any Python
dict.
This Python representation of a (collection of) JSON object(s) (2) is
translated into an actual instance of a class defined in this module.
This is what was most of the code in the classes inheriting from
JSONEncoder/JSONDecoder was about. These classes will be removed soon.

This commit needs testing.