fix(DampedSPM): damping functs are multidim
Implement the fact that the dumping functions are multivariate in the
coordinates.
The attribute self._damp is a list of Nd-long lists, each item being None or
a DampingFunction instance which acts on one of the Nd coordinates.
If None, no damping is applied along the corresponding coordinate.

Redefine proper set/reset/get-ters.
The __call__() method is corrected for proper computation of the
damping.

TODO:
* Fix export/import to/from JSON
* Test behaviour