refactor(Diabatizer): comp cost using data members
Cache the keys, origins and weights (in flat form) instead of passing
them as extra arguments in the cost function.
This makes better sense as they are not going to change throughout the
optimization procedure.
It also allows the formulation of a callback function for verbose
output in the future.