refactor(Diabatizer): use SymPolyMat.update
Use SymPolyMat.update method from Wout instead of reconstructing a new
SymPolyMat from scratch at every iteration.
This allows to modify only the relevant data of Wout, be it a bare
SymPolyMat of a Damped version (i.e. the update is type independent).
So, e.g. the x0s and damping functions remain the same.